# openapi-spec/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: cloud Job Scheduling API
  version: 1.0.0
paths:
...
  /workers/heartbeat:
    post:
      summary: worker heartbeat ping
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
      responses:
        "200":
          description: ok
        "404":
          description: worker not found